update Dockerfile to handle building versions
This commit adds a `version` build argument for the config to supply a specific version rather than always defaulting to the latest Closure Compiler release.  To match the URI format of the dl.google.com packages, the `version` build argument allows either a YYYYMMDD release version (e.g. `20200517`) or simply `latest`.  This commit does not expect to handle versions prepended with a 'v', e.g. `v20200517`.

The default value of `"latest"` for `version` allows the Cloud Build config to optionally exclude the `--build-arg` flag and `substitutions` blocks when building the image.  This may be an unnecessary addition but potentially offers some flexibility in modifying the build config.